Servers are Case Sensitive - OS X isn't. Missing end line fix.
Great plug-in. Three minor errors corrected.

Apple uses @2x for iPhone development not @2X.

Retina JS link needs to be RetinaImages/ not retinaImage/ to work on
standard servers. It was working on Mac localhost. Took ages to find
what was stopping it all working.

SS3 reported an end line missing.